factor out pointers test
